load books from backend

+++ backend/src/main/java/org/example/backend/types/Book.java
@@ -4,6 +4,6 @@ import lombok.With;
 import org.springframework.data.annotation.Id;
 
 @With
-public record Book( String title, String author, @Id int isbn, String summary, String image, int totalAmount,
+public record Book( String title, String author, @Id long isbn, String summary, String image, int totalAmount,
                     int totalBookedAmount ) {
 }
